PHPC-2348 Remove WriteException and move getWriteResult to BulkWriteException
