+++ src/MongoDB/Exception/BulkWriteException.stub.php
@@ -7,6 +7,10 @@
 
 namespace MongoDB\Driver\Exception;
 
-class BulkWriteException extends WriteException
+class BulkWriteException extends RuntimeException
 {
+    /** @var \MongoDB\Driver\WriteResult */
+    protected $writeResult;
+
+    final public function getWriteResult(): \MongoDB\Driver\WriteResult {}
 }
